Moved Function Related Methods
Moved methods related to functions in ExpressionContext to its own partial class
